Adjust the GUI lib with rules from linters and formatters
- Adjusted the GUI lib with rules from linters and formatters
- Organizing better the 'gui_lib' folder
- Created the 'widget_lib' folder
