+++ gui_lib/label.py
@@ -1,19 +1,10 @@
+"""This file has a set of classes related to "QtWidgets.QLabel"."""
+
 from PyQt5 import QtCore, QtWidgets
 
 
 class StandardLabel(QtWidgets.QLabel):
-
-    """
-    This class is used to create a Standard Label inheriting the "QtWidgets.QLabel" class.
-
-    Arguments:
-    - CentralWidget: the widget where the Label will be placed
-    - title: the text on the Label
-    - coordinate_X: the window X coordinate where the Label will be placed
-    - coordinate_Y: the window Y coordinate where the Label will be placed
-    - width: the width of the Label
-    - height: the height of the Label
-    """
+    """Class used to create a StandardLabel with "QtWidgets.QLabel"."""
 
     # Contants related to the push button
     DEFAULT_WIDTH = 200
@@ -28,6 +19,23 @@ class StandardLabel(QtWidgets.QLabel):
         width=DEFAULT_WIDTH,
         height=DEFAULT_HEIGHT,
     ):
+        """Create a StandardLabel object rom "QtWidgets.QLabel".
+
+        Arguments:
+        - CentralWidget: the widget where the Label will be placed
+        - title: the text on the Label
+        - coordinate_X: the window X coordinate inside the widget
+        - coordinate_Y: the window Y coordinate inside the widget
+        - width: the width of the Label
+        - height: the height of the Label
+        """
         super().__init__(CentralWidget)
-        self.setGeometry(QtCore.QRect(coordinate_X, coordinate_Y, width, height))
+        self.setGeometry(
+            QtCore.QRect(
+                coordinate_X,
+                coordinate_Y,
+                width,
+                height,
+            )
+        )
         self.setText(title)
