+++ gui_lib/pushbutton.py
@@ -1,20 +1,10 @@
+"""This file has a set of classes related to "QtWidgets.QPushButton"."""
+
 from PyQt5 import QtCore, QtWidgets
 
 
 class StandardPushButton(QtWidgets.QPushButton):
-
-    """
-    This class is used to create a Standard PushButton inheriting the "QtWidgets.QPushButton" class.
-
-    Arguments:
-    - CentralWidget: the widget where the PushButton will be placed
-    - title: the text on the PushButton
-    - coordinate_X: the window X coordinate where the PushButton will be placed
-    - coordinate_Y: the window Y coordinate where the PushButton will be placed
-    - width: the width of the PushButton
-    - height: the height of the PushButton
-    - onClickMethod: the callback method of the "onClick" PushButton event
-    """
+    """Class to create a StandardPushButton with "QtWidgets.QPushButton"."""
 
     # Contants related to the PushButton
     DEFAULT_WIDTH = 200
@@ -30,8 +20,27 @@ class StandardPushButton(QtWidgets.QPushButton):
         height=DEFAULT_HEIGHT,
         onClickMethod=None,
     ):
+        """
+        Create an StandardPushButton object from "QtWidgets.QPushButton".
+
+        Arguments:
+        - CentralWidget: the widget where the PushButton will be placed
+        - title: the text on the PushButton
+        - coordinate_X: the window X coordinate inside the widget
+        - coordinate_Y: the window Y coordinate inside the widget
+        - width: the width of the PushButton
+        - height: the height of the PushButton
+        - onClickMethod: the callback method of the "onClick" PushButton event
+        """
         super().__init__(CentralWidget)
-        self.setGeometry(QtCore.QRect(coordinate_X, coordinate_Y, width, height))
+        self.setGeometry(
+            QtCore.QRect(
+                coordinate_X,
+                coordinate_Y,
+                width,
+                height,
+            )
+        )
         self.setText(title)
         if onClickMethod:
             self.clicked.connect(onClickMethod)
