+++ gui_lib/textedit.py
@@ -1,18 +1,10 @@
+"""This file has a set of classes related to "QtWidgets.QTextEdit"."""
+
 from PyQt5 import QtCore, QtWidgets
 
 
 class StandardTextEdit(QtWidgets.QTextEdit):
-
-    """
-    This class is used to create a Standard TextEdit inheriting the "QtWidgets.QTextEdit" class.
-
-    Arguments:
-    - CentralWidget: the widget where the TextEdit will be placed
-    - coordinate_X: the window X coordinate where the TextEdit will be placed
-    - coordinate_Y: the window Y coordinate where the TextEdit will be placed
-    - width: the width of the TextEdit
-    - height: the height of the TextEdit
-    """
+    """Class used to create a StandardTextEdit with "QtWidgets.QTextEdit"."""
 
     # Contants related to the TextEdit
     DEFAULT_WIDTH = 200
@@ -26,5 +18,22 @@ class StandardTextEdit(QtWidgets.QTextEdit):
         width=DEFAULT_WIDTH,
         height=DEFAULT_HEIGHT,
     ):
+        """
+        Create an StandardTextEdit object from "QtWidgets.QTextEdit".
+
+        Arguments:
+        - CentralWidget: the widget where the TextEdit will be placed
+        - coordinate_X: the window X coordinate inside the widget
+        - coordinate_Y: the window Y coordinate inside the widget
+        - width: the width of the TextEdit
+        - height: the height of the TextEdit
+        """
         super().__init__(CentralWidget)
-        self.setGeometry(QtCore.QRect(coordinate_X, coordinate_Y, width, height))
+        self.setGeometry(
+            QtCore.QRect(
+                coordinate_X,
+                coordinate_Y,
+                width,
+                height,
+            )
+        )
