+++ gui_lib/window.py
@@ -1,13 +1,10 @@
+"""This file has a set of classes related to "QtWidgets.QMainWindow"."""
+
 from PyQt5 import QtWidgets
 
 
 class Window(QtWidgets.QMainWindow):
-    """
-    This class provides methods and attributes to create a window.
-
-    Arguments:
-    - window_title: the title of the window
-    """
+    """Class used to create a Window with "QtWidgets.QMainWindow"."""
 
     # Contants related to the window
     DEFAULT_WIDTH = 1200
@@ -15,13 +12,15 @@ class Window(QtWidgets.QMainWindow):
     DEFAULT_BORDER_SIZE = 20
 
     def __init__(self, window_title):
+        """
+        Create a window object from "QtWidgets.QMainWindow".
+
+        Arguments:
+        - window_title: the title of the window
+        """
         super().__init__()
         self.__setupWindow(window_title)
         self.__setDefaultWindowSize()
-
-    """
-    Private methods
-    """
 
     def __setupWindow(self, window_title):
         self.CentralWidget = QtWidgets.QWidget()
@@ -32,14 +31,11 @@ class Window(QtWidgets.QMainWindow):
         self.CentralWidget.resize(Window.DEFAULT_WIDTH, Window.DEFAULT_HEIGHT)
         self.resize(Window.DEFAULT_WIDTH, Window.DEFAULT_HEIGHT)
 
-    """
-    Public methods
-    """
-
     def getCentralWidget(self):
         """
-        Returns the central widget of the window.
+        Return the central widget of the window.
 
-        All GUI components (buttons, checkboxes, tables, etc.) must have this widget as argument.
+        All GUI components (buttons, checkboxes, tables, etc.) must have this
+        widget as argument.
         """
         return self.CentralWidget
